Update to latest SQIsign implementation

+++ crypto_sign/sqisign3/mikes_arithmetic/rng.h
@@ -3,6 +3,8 @@
 #ifndef rng_h
 #define rng_h
 
+#include <stddef.h>
+
 /**
  * Randombytes initialization.
  * Initialization may be needed for some random number generators (e.g. CTR-DRBG).
@@ -23,6 +25,6 @@ void randombytes_init(unsigned char *entropy_input,
  * @param[in] xlen Number of random bytes to be generated
  * @return int 0 on success, -1 otherwise
  */
-int randombytes(unsigned char *x, unsigned long long xlen);
+int randombytes(unsigned char *x, size_t xlen);
 
 #endif /* rng_h */
